custom default configs

+++ init.sh
@@ -8,7 +8,11 @@ source cbuild/functions.sh
 #config
 if [ ! -f ".config" ]; then
     print "${YELLOW}./.config doesn't exist\n"
-    cp cbuild/default.config.sh .config
+    if [ -f "default.config.sh" ]; then
+        cp default.config.sh .config
+    else
+        cp cbuild/default.config.sh .config
+    fi
     print "${YELLOW}Default config created.\nEdit it's values.\n"
     exit
 fi
